chore: split registration endpoints into separate files
